requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aioprocessing==2.0.1
aiosignal==1.3.1
alembic==1.13.1
analytics-python==1.2.9
annotated-types==0.6.0
anyio==4.4.0
async-timeout==4.0.3
attrs==22.1.0
aws-cdk-lib==2.134.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.1
aws-cdk.aws-lambda-python-alpha==2.134.0a0
backoff==2.2.1
boto3==1.34.113
botocore==1.34.113
cattrs==22.1.0
certifi==2024.7.4
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.7
colorama==0.4.6
constructs==10.3.0
coverage==6.5.0
cryptography==42.0.7
datadog-cdk-constructs-v2==1.15.0
debugpy==1.8.1
distro==1.9.0
docker-pycreds==0.4.0
ecdsa==0.19.0
emoji==2.12.1
exceptiongroup==1.2.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
gql==3.5.0
graphql-core==3.2.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
iniconfig==1.1.1
janus==1.0.0
Jinja2==3.1.4
jiter==0.5.0
jmespath==1.0.1
jsii==1.102.0
langfuse==2.44.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.1
mdurl==0.1.2
monotonic==1.6
moto==5.0.8
multidict==6.0.5
numpy==1.24.4
openai==1.40.6
packaging==23.2
pgvector==0.2.5
pip-review==1.3.0
platformdirs==4.2.2
pluggy==1.5.0
promise==2.3
protobuf==4.25.3
psutil==5.9.8
psycopg2-binary==2.9.9
publication==0.0.3
pyarrow==16.1.0
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
Pygments==2.18.0
pyparsing==3.0.9
pytest==8.3.2
pytest-cov==4.0.0
python-dateutil==2.8.2
python-json-logger==2.0.7
pytz==2022.5
PyYAML==6.0.1
regex==2023.12.25
requests==2.32.2
requests-toolbelt==1.0.0
responses==0.22.0
rich==13.7.1
Rx==1.6.3
s3transfer==0.10.1
sentry-sdk==2.13.0
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.32
sshpubkeys==3.3.1
structlog==24.4.0
tabulate==0.9.0
tenacity==8.3.0
tiktoken==0.6.0
toml==0.10.2
tomli==2.0.1
tqdm==4.66.4
typeguard==2.13.3
types-toml==0.10.8
typing_extensions==4.12.2
urllib3==1.26.19
uuid_utils==0.9.0
Werkzeug==3.0.3
wrapt==1.16.0
xmltodict==0.13.0
yarl==1.9.4
zipp==3.19.2
